Ajout du handler de post
- Correctif des fonctions BDD liés aux posts et commentaires pour postgres
- Correctif de la fonction CreateNewPost() dans le service PostAndComments pour récupérer l'ID de façon plus propre
- Implémentation basique du handler de nouveau post. Champs nécessaires dans le formulaire : title (facultatif), content, privacy, mode(caché, edit ou newpost), tags(pour l'instant séparés par des espaces)
- Le handler récupère l'ID de l'utilisateur connecté dans les cookies, adaptation de la fonction editPost pour ne pouvoir modifier que ses propres messages (check sécurité)

+++ backend/internal/service/postAndCommentsService .go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"fmt"
 	"social-network/backend/internal/model"
 	"social-network/backend/internal/repository"
@@ -23,12 +24,7 @@ func NewPostAndCommentsService(pr *repository.PostRepo, tr *repository.TagRepo, 
 * Paramètres : ID du posteur et postData (titre, contenu, confidentialité, tags)
 */
 func (s *PostAndCommentsService) CreateNewPost(authorID string, postData model.Post) error {
-	err := s.postRepo.CreateNewPost(authorID, postData)
-	if err != nil {
-		return err
-	}
-	
-	postID, err := s.postRepo.GetPostIDFromContent(authorID, postData.Content)
+	postID, err := s.postRepo.CreateNewPost(authorID, postData)
 	if err != nil {
 		return err
 	}
@@ -99,14 +95,28 @@ func (s *PostAndCommentsService) DeleteComment(commentID int, mode string) error
 * 2. Vérifie si des tags ont été ajoutés et les ajoute si c'est le cas
 * Paramètres : model.Post complet 
 */
-func (s *PostAndCommentsService) EditPost(postData model.Post) error {
-	err := s.postRepo.UpdateExistingPost(postData)
+func (s *PostAndCommentsService) EditPost(authorID string, postData model.Post) error {
+	posterID, err := s.postRepo.GetPostAuthorID(postData.ID)
+
+	if posterID != authorID {
+		return errors.New("utilisateur non autorisé")
+	}
+
+
+	err = s.postRepo.UpdateExistingPost(postData)
 	if err != nil {
 		return err
 	}
 
-	s.tagRepo.DeletePostTags(postData.ID)
-	s.tagRepo.AddPostTags(postData.ID, postData.Tags)
+	err = s.tagRepo.DeletePostTags(postData.ID)
+	if err != nil {
+		return err
+	}
+
+	err = s.tagRepo.AddPostTags(postData.ID, postData.Tags)
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
